refactor: Reorganizing the `types` folder and file structure
