+++ src/frss-extension/elements/EvidenceSource/properties.ts
@@ -1,6 +1,6 @@
 // import the prefix
 import { bpmn4frssPrefix } from '../../common';
-import { RenderableElementWithIconProps } from '../../types/props';
+import { RenderableElementWithIconProps } from '../../types/properties/properties';
 import icon
   from './assets/evidence-source.png';
