--- src/frss-extension/elements/template/description.md
@@ -1,23 +1,0 @@
-# `x`
-
-Some general description.
-
-Properties of the `x`:
-
-- [x] `propertyName`
-
-  Property description.
-
-## Implemented features
-
-- [x] feature name
-
-_optional:_ some comment about the implemented features.
-
-## To be implemented
-
-- [ ] feature that was not implemented yet.
-  - [x] sub-feature that contributes to implementation of the parent feature and that is implemented already
-  - [ ] sub-feature that contributes to implementation of the parent feature and that has not been implemented yet
-
-Comment about the state of implementation of the element.
