--- src/frss-extension/elements/template/index.ts
@@ -1,19 +1,0 @@
-import FrssElement from '../../types';
-// import controls from './controls';
-import definition from './definition';
-// import icon from './assets/x.png';
-import properties from './properties';
-// import render from './rendererEntry';
-
-const element: FrssElement = {
-  // controls,
-  definition,
-  // icon,
-  properties,
-  // render,
-};
-
-/**
- * The default export exports all properties of a custom element.
- */
-export default element;
