+++ src/frss-extension/types/properties/properties.ts
