+++ src/frss-extension/types/renderer/rendererEntry.ts
