+++ src/frss-extension/types/rules/connection.ts
@@ -1,0 +1,13 @@
+export interface HasConnectionRule {
+  connectionRule: (source: any, target: any) => (boolean
+  | { type: string } | void),
+  shouldCheckConnection: (source: any, target: any) => boolean,
+}
+
+export const hasConnectionRule = (rules: any):
+rules is HasConnectionRule => {
+  const checkRule = rules as HasConnectionRule;
+
+  return checkRule.connectionRule !== undefined
+    && checkRule.shouldCheckConnection !== undefined;
+};
