+++ src/frss-extension/types/rules/events/preDelete.ts
@@ -1,0 +1,16 @@
+export type HasPreDeleteEvents = {
+  preDeleteEvents: PreDeleteEvent[],
+};
+
+export interface PreDeleteEvent {
+  preDeleteEvent: (event: any) => void,
+  shouldTriggerPreDelete: (event: any) => boolean,
+}
+
+export const hasPreDeleteEvent = (rules: any):
+rules is HasPreDeleteEvents => {
+  const checkRule = rules as HasPreDeleteEvents;
+
+  return checkRule.preDeleteEvents !== undefined
+    && checkRule.preDeleteEvents.length > 0;
+};
